BREAKING: Update to support Dart 3 & Flutter 3.10.0
This patch updates the screenshot system to use
FlutterView instead of Window.

This patch creates a new function and deletes the old one.
Context is now required as it is needed to lookup the current
view.

This is a breaking change but is required to go forward since
use of Window is deprecated.

https://api.flutter.dev/flutter/dart-ui/window.html
https://api.flutter.dev/flutter/widgets/View/of.html